requirements.txt
tensorflow
numpy
pillow
scikit-learn
opencv-python
matplotlib
scikit-image
python-dotenv
streamlit